Refactor shell-related startup scripts.
This change cleans up the structure of the zsh and bash startup scripts
and, more significantly, incorporates a number of improvements from
PimpMyTerminal (https://drasite.com/blog/Pimp%20my%20terminal)

This change also adds the zsh-autosuggestion plugin and changes the
prompt to Starship (https://starship.rs). Starship must be installed
beforehand.

Though, there should be a fallback in the case where Starship cannot be
installed on the system, which will be implemented in the next change.